% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-filter.R
\name{ex_filter_name_value}
\alias{ex_filter_name_value}
\title{filter on name and value}
\usage{
ex_filter_name_value(.data, var_name, value)
}
\arguments{
\item{.data}{`data.frame` to be filtered}

\item{var_name}{`character`, name of column within `.data` to filter}

\item{value}{object,}
}
\value{
filtered copy of `.data`
}
\description{
filter on name and value
}
\examples{
ex_filter_name_value(mtcars, "cyl", 6)
}
